write user location to local db

+++ app/src/main/java/com/example/tikimvvm/db/entity/UserLocation.kt
@@ -16,7 +16,7 @@ data class UserLocation(
 
         val address: String,
 
-        val numOfSatellite: Double,
+        val numOfSatellite: Int,
 
         val hdop: Double,
 )
